feat/mobile: add transaction filtering logic
